+++ src/main/scala/edu/ie3/simona/service/results/ResultServiceProxy.scala
@@ -27,7 +27,7 @@ import edu.ie3.simona.service.results.Transformer3wResultSupport.{
   Transformer3wKey,
 }
 import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
-import edu.ie3.simona.util.TickUtil.RichZonedDateTime
+import edu.ie3.simona.util.TickUtil.toTick
 import org.apache.pekko.actor.typed.scaladsl.{Behaviors, StashBuffer}
 import org.apache.pekko.actor.typed.{ActorRef, Behavior, PostStop}
 import org.slf4j.Logger