+++ src/main/scala/edu/ie3/util/scala/quantities/SquantsUtils.scala
@@ -18,18 +18,18 @@ import squants.time.Frequency
 import scala.math.Pi
 
 object SquantsUtils {
-  implicit class RichEnergy(energy: Energy) {
+  extension (energy: Energy) {
     def calcVolume(that: EnergyDensity): Volume = CubicMeters(
       energy.toKilowattHours / that.toKilowattHoursPerCubicMeter
     )
   }
-  implicit class RichPower(power: squants.Power) {
+  extension (power: squants.Power) {
     def /(that: ReactivePower): Dimensionless = Each(
       power.toWatts / that.toVars
     )
   }
 
-  implicit class RichElectricPotential(
+  extension (
       electricPotential: ElectricPotential
   ) {
     def multiplyWithDimensionless(
@@ -39,16 +39,16 @@ object SquantsUtils {
     )
   }
 
-  implicit class RichThermalCapacity(
+  extension (
       thermalCapacity: ThermalCapacity
   ) {
     def toWattHoursPerKelvin: Double =
-      this.thermalCapacity.toJoulesPerKelvin / 3600
+      thermalCapacity.toJoulesPerKelvin / 3600
     def toWattSecondsPerKelvin: Double =
-      this.thermalCapacity.toJoulesPerKelvin // Joule == Ws
+      thermalCapacity.toJoulesPerKelvin // Joule == Ws
   }
 
-  implicit class RichIrradiance(
+  extension (
       irradiance: Irradiance
   ) {
     def *(that: Time): Irradiation = WattHoursPerSquareMeter(